+++ tests/test_delete.py
@@ -37,7 +37,9 @@ def test_errors(httpserver, code):
 
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SCIMClient(client, resource_types=(User,))
-    response = scim_client.delete(User, "2819c223-7f76-453a-919d-413861904646")
+    response = scim_client.delete(
+        User, "2819c223-7f76-453a-919d-413861904646", raise_scim_errors=False
+    )
 
     assert response == Error(
         schemas=["urn:ietf:params:scim:api:messages:2.0:Error"],